[Add] UserFactory 유저 삭제 기능
jest에서 테스트 파일 간의 독립성이 보장되지 않아 한 인스턴스를 테스트 파일 간에 공유하면 중복되는 유저를 생성하려고 해
테스트와 관계없는 에러가 발생하였다.

따라서 인스턴스가 아닌 class를 export하고 각 인스턴스를 각 테스트 파일에서 사용하고, 그 파일에서 생성한 유저들만 삭제하도록 수정하였다.
